refactor | todo 추가 로직 hooks 로 리펙터링
